requirements-pypi.txt
torch
transformers==4.46.3
ipykernel
tqdm
python-dotenv
h5py
scikit-learn
plotly
pandas
matplotlib
pycocotools
nbformat
speaksee
munkres
open_clip_torch
sentence_transformers
nltk